agent/requirements.txt
# Agent dependencies
google-adk
requests
streamlit